fix: products 관련 api 변경 및 요청 사항 수정

+++ src/main/java/com/freeder/buclserver/domain/productcategory/dto/ProductCategoryDTO.java
@@ -14,5 +14,5 @@ public class ProductCategoryDTO {
 	private int reward;
 	private int discountRate;
 	private int totalReviewCount;
-	private double averageRating;
+	private float averageRating;
 }
